avoid string operations on numbers
multiple methods use regex to check if input argument is a number, this
is rather inefficient and to add to that if it is not a number, these
methods simply return unchanged input value without indicating an error
in any way.

_round_to_precision method in addition to that converts $precision value
to a string for the sake of rounding it, prepends it with "-" and then
parses the result back to number. This logic is hard to understand and
it is also not very efficient.